+++ internal/handlers/take_backup_test.go
@@ -307,9 +307,9 @@ func TestTBOperationHandlerRunningOperationCompletedSuccessfully(t *testing.T) {
 	assert.Equal(t, Ydb.StatusIds_NOT_FOUND, ydbOpStatus.GetOperation().GetStatus())
 
 	// check metrics
-	assert.Equal(t, int64(450), mon.GetMetrics()["storage_bytes_written"])
-	assert.Equal(t, int64(1), mon.GetMetrics()["backups_succeeded_count"])
-	assert.Equal(t, int64(0), mon.GetMetrics()["backups_failed_count"])
+	assert.Equal(t, float64(450), mon.GetMetrics()["storage_bytes_written"])
+	assert.Equal(t, float64(1), mon.GetMetrics()["backups_succeeded_count"])
+	assert.Equal(t, float64(0), mon.GetMetrics()["backups_failed_count"])
 }
 
 func TestTBOperationHandlerRunningOperationCancelled(t *testing.T) {
@@ -384,9 +384,9 @@ func TestTBOperationHandlerRunningOperationCancelled(t *testing.T) {
 	assert.Equal(t, Ydb.StatusIds_NOT_FOUND, ydbOpStatus.GetOperation().GetStatus())
 
 	// check metrics
-	assert.Equal(t, int64(0), mon.GetMetrics()["storage_bytes_written"])
-	assert.Equal(t, int64(0), mon.GetMetrics()["backups_succeeded_count"])
-	assert.Equal(t, int64(1), mon.GetMetrics()["backups_failed_count"])
+	assert.Equal(t, float64(0), mon.GetMetrics()["storage_bytes_written"])
+	assert.Equal(t, float64(0), mon.GetMetrics()["backups_succeeded_count"])
+	assert.Equal(t, float64(1), mon.GetMetrics()["backups_failed_count"])
 }
 
 func TestTBOperationHandlerDeadlineExceededForCancellingOperation(t *testing.T) {
@@ -463,9 +463,9 @@ func TestTBOperationHandlerDeadlineExceededForCancellingOperation(t *testing.T) 
 	assert.Equal(t, false, ydbOpStatus.GetOperation().GetReady())
 
 	// check metrics
-	assert.Equal(t, int64(0), mon.GetMetrics()["storage_bytes_written"])
-	assert.Equal(t, int64(0), mon.GetMetrics()["backups_succeeded_count"])
-	assert.Equal(t, int64(1), mon.GetMetrics()["backups_failed_count"])
+	assert.Equal(t, float64(0), mon.GetMetrics()["storage_bytes_written"])
+	assert.Equal(t, float64(0), mon.GetMetrics()["backups_succeeded_count"])
+	assert.Equal(t, float64(1), mon.GetMetrics()["backups_failed_count"])
 }
 
 func TestTBOperationHandlerCancellingOperationInProgress(t *testing.T) {
@@ -630,9 +630,9 @@ func TestTBOperationHandlerCancellingOperationCompletedSuccessfully(t *testing.T
 	assert.Equal(t, Ydb.StatusIds_NOT_FOUND, ydbOpStatus.GetOperation().GetStatus())
 
 	// check metrics
-	assert.Equal(t, int64(450), mon.GetMetrics()["storage_bytes_written"])
-	assert.Equal(t, int64(1), mon.GetMetrics()["backups_succeeded_count"])
-	assert.Equal(t, int64(0), mon.GetMetrics()["backups_failed_count"])
+	assert.Equal(t, float64(450), mon.GetMetrics()["storage_bytes_written"])
+	assert.Equal(t, float64(1), mon.GetMetrics()["backups_succeeded_count"])
+	assert.Equal(t, float64(0), mon.GetMetrics()["backups_failed_count"])
 }
 
 func TestTBOperationHandlerCancellingOperationCancelled(t *testing.T) {
@@ -707,9 +707,9 @@ func TestTBOperationHandlerCancellingOperationCancelled(t *testing.T) {
 	assert.Equal(t, Ydb.StatusIds_NOT_FOUND, ydbOpStatus.GetOperation().GetStatus())
 
 	// check metrics
-	assert.Equal(t, int64(0), mon.GetMetrics()["storage_bytes_written"])
-	assert.Equal(t, int64(0), mon.GetMetrics()["backups_succeeded_count"])
-	assert.Equal(t, int64(1), mon.GetMetrics()["backups_failed_count"])
+	assert.Equal(t, float64(0), mon.GetMetrics()["storage_bytes_written"])
+	assert.Equal(t, float64(0), mon.GetMetrics()["backups_succeeded_count"])
+	assert.Equal(t, float64(1), mon.GetMetrics()["backups_failed_count"])
 }
 
 func TestTBOperationHandlerCancellingOperationCancelledWithRemovingDataFromS3(t *testing.T) {
@@ -804,10 +804,10 @@ func TestTBOperationHandlerCancellingOperationCancelledWithRemovingDataFromS3(t 
 	assert.Equal(t, 0, len(objects))
 
 	// check metrics
-	assert.Equal(t, int64(450), mon.GetMetrics()["storage_bytes_written"])
-	assert.Equal(t, int64(450), mon.GetMetrics()["storage_bytes_deleted"])
-	assert.Equal(t, int64(0), mon.GetMetrics()["backups_succeeded_count"])
-	assert.Equal(t, int64(1), mon.GetMetrics()["backups_failed_count"])
+	assert.Equal(t, float64(450), mon.GetMetrics()["storage_bytes_written"])
+	assert.Equal(t, float64(450), mon.GetMetrics()["storage_bytes_deleted"])
+	assert.Equal(t, float64(0), mon.GetMetrics()["backups_succeeded_count"])
+	assert.Equal(t, float64(1), mon.GetMetrics()["backups_failed_count"])
 }
 
 func TestTBOperationHandlerRetriableErrorForRunningOperation(t *testing.T) {
